fix: make css process clearly on hmr
